Declare some integers as unsigned. (#42)
This fixes some gcc warnings of comparing signed with unsigned
integers. We should take that serious. Comparing a negative integer
to a positive unsigned integer must never happen.

+++ src/Methods/FiniteTempString.cpp
@@ -268,7 +268,7 @@ namespace SSAGES
 	void FiniteTempString::StringUpdate()
 	{
 		size_t i;
-		int centersize = _centers.size();
+		size_t centersize = _centers.size();
 		double alpha_star;
 		std::vector<double> alpha_starv;
 		std::vector<double> cvs_new;
@@ -285,7 +285,7 @@ namespace SSAGES
 			sendneighbor = 1;
 			recvneighbor = _world.size()-1;
 		} 
-		else if (_mpiid == _world.size()-1){
+		else if (_mpiid == (unsigned int)_world.size() - 1){
 			sendneighbor = 0;
 			recvneighbor = _world.size() - 2;
 		}
